Refactoring the about page
