+++ public/index.ts
@@ -1,0 +1,15 @@
+import brazil from './brazil.png'
+import spain from './spain.png'
+import bulgaria from './bulgaria.png'
+import netherlands from './netherlands.png'
+import briefcase from './briefcase.png'
+import graduation from './graduation.png'
+import world from './world.png'
+
+export const brazilFlag = brazil;
+export const spainFlag = spain;
+export const bulgariaFlag = bulgaria;
+export const netherlandsFlag = netherlands;
+export const briefcaseIcon = briefcase;
+export const graduationIcon = graduation;
+export const worldIcon = world;
