+++ src/components/About/components/LinkButton/styles.ts
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const LinkButtonContainer = styled.div`
+  min-width: 15rem;
+  padding: 1rem;
+  border-radius: 8px;
+  background: #5CA0C5;
+
+  color: #17161A;
+  font-weight: bold;
+`
